fix possible loss of fraction

+++ Observer/Observers/MeasurementStatistics.cs
@@ -2,19 +2,19 @@
 
 public class MeasurementStatistics
 {
-    public int TemperatureAverage { get; set; }
+    public double TemperatureAverage { get; set; }
 
     public int TemperatureMin { get; set; }
 
     public int TemperatureMax { get; set; }
 
-    public int HumidityAverage { get; set; }
+    public double HumidityAverage { get; set; }
 
     public int HumidityMin { get; set; }
 
     public int HumidityMax { get; set; }
 
-    public int PressureAverage { get; set; }
+    public double PressureAverage { get; set; }
 
     public int PressureMin { get; set; }
